VideoClustering_2/VideoClusteringProcessing/requirement.txt
#####################
sklearn
webcolors
imutils
kafka-python

####### Model #######
scikit-learn==1.0.2
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.41.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
thop 
torch==1.9.1
torchvision==0.10.1
torchaudio==0.9.1
faiss-cpu
